materials added

+++ includes/rtv1.h
@@ -64,13 +64,13 @@ char					**get_kernel_source(t_cl *cl, char *path);
  ** ---rendering---
  */
 
-int					*render(t_cl *cl, int w, int h);
+float					*render(t_cl *cl, int w, int h);
 
 /*
  ** ---coloring---
  */
 
-void					update_texture(SDL_Texture *texture, int width, int height, int *r);
+void					update_texture(SDL_Texture *texture, int width, int height, float *r);
 
 /*
  ** ---controls---
